requirements.txt
black==22.12.0
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
idna==3.4
fastapi==0.89.1
mypy-extensions==0.4.3
pathspec==0.11.0
platformdirs==2.6.2
pydantic==1.10.4
python-dotenv==0.21.1
requests==2.28.2
tomli==2.0.1
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
tzdata==2022.7